requirements.txt
git+https://github.com/django/django.git#Django
Markdown==2.4
Pillow==2.4.0
argparse==1.2.1
crispy-forms-foundation==0.3.3
django-allauth==0.16.1
django-appconf==0.6
django-autocomplete-light==2.0.0a17
django-autoslug==1.7.2
django-avatar==2.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.2
django-markupfield==1.2.0
django-postman==3.1.0
django-sendmail-backend==0.1.2
django-sitetree==1.0.0
django-zurb-foundation==5.1.1
easy-thumbnails==2.0.1
html2text==2014.4.5
oauthlib==0.6.1
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
